remove tooling, inline Debug, fix publishing
